# Improve performance of `ChatChannel` database model conversions ~7 times