--- a/Sources/StreamChat/Database/DTOs/MemberModelDTO.swift
+++ b/Sources/StreamChat/Database/DTOs/MemberModelDTO.swift
@@ -92,22 +92,6 @@ extension MemberDTO {
         new.id = memberId
         return new
     }
-
-    static func lastActiveMembersRequest(cid: ChannelId, context: NSManagedObjectContext) -> NSFetchRequest<MemberDTO> {
-        let request = NSFetchRequest<MemberDTO>(entityName: MemberDTO.entityName)
-        request.predicate = NSPredicate(format: "channel.cid == %@", cid.rawValue)
-        request.sortDescriptors = [
-            ChannelMemberListSortingKey.lastActiveSortDescriptor,
-            ChannelMemberListSortingKey.defaultSortDescriptor
-        ]
-        request.fetchLimit = context.localCachingSettings?.chatChannel.lastActiveMembersLimit ?? 100
-        return request
-    }
-    
-    static func loadLastActiveMembers(cid: ChannelId, context: NSManagedObjectContext) -> [MemberDTO] {
-        let request = lastActiveMembersRequest(cid: cid, context: context)
-        return load(by: request, context: context)
-    }
 }
 
 extension NSManagedObjectContext {