--- a/Sources/StreamChat/Database/DTOs/MessageDTO.swift
+++ b/Sources/StreamChat/Database/DTOs/MessageDTO.swift
@@ -1323,21 +1323,28 @@ private extension ChatMessage {
 
         if let currentUser = context.currentUser {
             isSentByCurrentUser = currentUser.user.id == dto.user.id
-            currentUserReactions = Set(
-                MessageReactionDTO
-                    .loadReactions(ids: dto.ownReactions, context: context)
-                    .compactMap { try? $0.asModel() }
-            )
+            if !dto.ownReactions.isEmpty {
+                currentUserReactions = Set(
+                    MessageReactionDTO
+                        .loadReactions(ids: dto.ownReactions, context: context)
+                        .compactMap { try? $0.asModel() }
+                )
+            } else {
+                currentUserReactions = []
+            }
         } else {
             isSentByCurrentUser = false
             currentUserReactions = []
         }
 
-        latestReactions = Set(
-            MessageReactionDTO
-                .loadReactions(ids: dto.latestReactions, context: context)
-                .compactMap { try? $0.asModel() }
-        )
+        latestReactions = {
+            guard !dto.latestReactions.isEmpty else { return Set() }
+            return Set(
+                MessageReactionDTO
+                    .loadReactions(ids: dto.latestReactions, context: context)
+                    .compactMap { try? $0.asModel() }
+            )
+        }()
 
         threadParticipants = dto.threadParticipants.array
             .compactMap { $0 as? UserDTO }
@@ -1351,7 +1358,7 @@ private extension ChatMessage {
             .sorted { $0.id.index < $1.id.index }
 
         latestReplies = {
-            guard !dto.replies.isEmpty else { return [] }
+            guard dto.replyCount > 0 else { return [] }
             return MessageDTO.loadReplies(for: dto.id, limit: 5, context: context)
                 .compactMap { try? ChatMessage(fromDTO: $0, depth: depth) }
         }()