--- a/Sources/StreamChat/Database/DTOs/UserDTO.swift
+++ b/Sources/StreamChat/Database/DTOs/UserDTO.swift
@@ -115,22 +115,6 @@ extension UserDTO {
         new.teams = []
         return new
     }
-
-    static func lastActiveWatchersRequest(cid: ChannelId, context: NSManagedObjectContext) -> NSFetchRequest<UserDTO> {
-        let request = NSFetchRequest<UserDTO>(entityName: UserDTO.entityName)
-        request.sortDescriptors = [
-            UserListSortingKey.lastActiveSortDescriptor,
-            UserListSortingKey.defaultSortDescriptor
-        ]
-        request.predicate = NSPredicate(format: "ANY watchedChannels.cid == %@", cid.rawValue)
-        request.fetchLimit = context.localCachingSettings?.chatChannel.lastActiveWatchersLimit ?? 100
-        return request
-    }
-    
-    static func loadLastActiveWatchers(cid: ChannelId, context: NSManagedObjectContext) -> [UserDTO] {
-        let request = lastActiveWatchersRequest(cid: cid, context: context)
-        return load(by: request, context: context)
-    }
 }
 
 extension NSManagedObjectContext: UserDatabaseSession {