… Sources/StreamChat/Query/Sorting/ChannelMemberListSortingKey.swift
… Sources/StreamChat/Query/Sorting/UserListSortingKey.swift
… Sources/StreamChat/Utils/Database/NSManagedObject+Extensions.swift
… Tests/StreamChatTests/Database/DTOs/ChannelDTO_Tests.swift