### 🔗 Issue Links

Related to [PBE-5004](https://stream-io.atlassian.net/browse/PBE-5004)

### 🎯 Goal

Improve performance of `ChatChannel` database model conversions by removing additional fetch requests

### 📝 Summary

- Roughly 7x CPU usage reduction when converting channel database models to immutable `ChatChannel`
- CPU time is saved by reducing additional fetch requests and using the DTOs data directly

### 🛠 Implementation

In #3299 we removed CoreDataLazy and this opened up an opportunity to optimise `asModel()` functions (these convert local database data to immutable models like `ChatChannel`). This PR replaces additional DB calls with accessing the NSManagedObject's data directly.

### 🎨 Showcase

Profiling steps: log out Chewbacca, log in with Chewbacca, wait until CPU usage settles. Observe the time spent in converting `ChannelDTO` to `ChatChannel` when the user is opening the channel list.

Measured: iPhone 12 Pro Max
279 ms > 38 ms 7x improvement

| Before | After |
| ------ | ----- |
|  <img width="697" alt="Baseline" src="https://github.com/user-attachments/assets/289a51a4-3691-4cb5-90ed-c3b7238d9097">   | <img width="581" alt="Final" src="https://github.com/user-attachments/assets/8e2bd975-3960-4124-903e-54dc61e2a2f8"> |

### 🧪 Manual Testing Notes

Exploratory testing around variety of channel actions, e.g. send a message > does everything look allright.

### ☑️ Contributor Checklist

- [x] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [x] This change should be manually QAed
- [x] Changelog is updated with client-facing changes
- [ ] Changelog is updated with new localization keys
- [ ] New code is covered by unit tests
- [ ] Comparison screenshots added for visual changes
- [ ] Affected documentation updated (docusaurus, tutorial, CMS)

[PBE-5004]: https://stream-io.atlassian.net/browse/PBE-5004?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ